pyproject.toml
[tool.poetry]
name = "siren_super_resolution"
version = "0.1.0"
description = "Siren application to super-resolution"
authors = ["ogvalt <ogvalt@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
torch = "^1.6.0"
torchvision = "^0.7.0"
python-dotenv = "^0.14.0"
hydra-core = "^0.11.3"
catalyst = "20.04.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
